+++ src/component/switch-sector.component.tsx
@@ -1,8 +1,5 @@
 import React from "react";
-import {
-  HighlightKeyCombination,
-  KeyLabel,
-} from "../model/device-layout.model";
+import { HighlightKeyCombination, KeyLabel } from "tangent-cc-lib";
 import { cos, sin } from "../util/math.util";
 import KeyLabelComponent from "./key-label.component";