+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Code refactor with tangent-cc-lib
+
 ## [4.0](https://github.com/andy23512/monkeytype-cc-extension/releases/tag/4.0)
 
 - Support fourth layer